The Sprockets library originally pulled from thingiverse and written by kresty.
I've added some chain profiles and made a few minor adjustments to the dimensioning of the sprocket output so as to better fit the chain profiles for which they were intended. Effectively I added a bit more tolerance to the chain widths as the perfectly dimensioned sprockets the original was spitting out were too perfect and needed filing.